requirements.txt
torch
torchvision
opencv-python
scikit-learn
tqdm
numpy
Pillow